fix: clear nightly clippy lints surfaced by the CI toolchain refresh
16b119a made CI `rustup update nightly` at container runtime; the newer
nightly clippy flags lints the stale baked toolchain did not, breaking
`cargo clippy --workspace -- -D warnings` (ci.yml) at rlvgl-core. Fix all
of them (clippy halts at the first erroring crate, so they surface in
layers):

- core/src/asset.rs: `and_then(|ch| if .. { Some(ch) } else { None })`
  -> `Option::filter` (manual_filter).
- core/src/plugins/fontdue.rs: drop redundant `metrics.ymin as i32`
  (ymin is already i32) (unnecessary_cast).
- widgets/src/span.rs: drop the no-op `.max(0)` on a `usize`.
- platform/src/serializer_arm.rs: index loop -> `iter_mut().skip(2)`
  (needless_range_loop; #[cfg(test)] synthesis).
- platform/tests/blend_row_override.rs: drop redundant `expected_r as i32`.
- examples/disco-sim/tests/typed_dma2d_host.rs: `drop()` of a non-Drop
  borrow token -> `let _ =` (dropping_copy_types); borrow still ends at
  the same statement.
- examples/disco-sim/tests/playit_automation.rs: `iter().any(|p| *p == x)`
  -> `contains(&x)`.

All clippy-mechanical equivalents — no behavior change. `cargo +nightly
clippy --workspace --all-targets -- -D warnings` is now clean; core 236 /
widgets 331 / serializer_arm unit tests pass.

+++ core/src/plugins/fontdue.rs
@@ -141,7 +141,7 @@ impl FontMetrics for FontdueFont<'_> {
         let font = get_cached_font(self.font_data);
         let metrics = font.metrics(ch, self.px);
         let advance_fp16 = round_to_i32(metrics.advance_width * 16.0).clamp(0, u16::MAX as i32);
-        let bearing_y = metrics.ymin as i32 + metrics.height as i32;
+        let bearing_y = metrics.ymin + metrics.height as i32;
         Some(GlyphInfo {
             advance_fp16: advance_fp16 as u16,
             bearing_x: metrics.xmin.clamp(i16::MIN as i32, i16::MAX as i32) as i16,
